Swap out openshift client for kubernetes client
All the functionality provided by the openshift client library has been
upstreamed to the kubernetes client library, so it's no longer necessary
for both of these.

This also switches community.kubernetes to kubernetes.core as the former
has been deprecated for some time now. This change only affects the
testing and is a drop in replacement.