Update ansible to version 2.15
This upgrades the test suite to use ansible 2.15. This update triggers a
number of other changes that have to be made. Molecule no longer handles
linting, so linting has to be handled by ansible-lint now. Running
ansible-lint generate quite a few linting issues. I've corrected enough
of the linting issues to get it to pass using the security profile. The
upgrade to ansible 2.15 also generated quited a few new sanity errors.
I've fixed all the documentation errors. Apparently, "message" is a
reserved module parameter, however. For now, I've added a sanity ignore
for this, which affects the k8s_event and k8s_status modules. This
parameter should probably be renamed, but this would obviously be a
breaking change.