Swap out openshift client for kubernetes client (#47)
* Swap out openshift client for kubernetes client

All the functionality provided by the openshift client library has been
upstreamed to the kubernetes client library, so it's no longer necessary
for both of these.

This also switches community.kubernetes to kubernetes.core as the former
has been deprecated for some time now. This change only affects the
testing and is a drop in replacement.

* Update ansible to version 2.15

This upgrades the test suite to use ansible 2.15. This update triggers a
number of other changes that have to be made. Molecule no longer handles
linting, so linting has to be handled by ansible-lint now. Running
ansible-lint generate quite a few linting issues. I've corrected enough
of the linting issues to get it to pass using the security profile. The
upgrade to ansible 2.15 also generated quited a few new sanity errors.
I've fixed all the documentation errors. Apparently, "message" is a
reserved module parameter, however. For now, I've added a sanity ignore
for this, which affects the k8s_event and k8s_status modules. This
parameter should probably be renamed, but this would obviously be a
breaking change.

* Fix ansible versions in github workflow